damage on dummy
created a dummy with hp. Each click generates a random number (damage) for the dummy from 100 to 150 units.
and created 2 buttons with which I will manipulate the number (damage) that will fall out when clicking. Analogue of untwisting and untwisting in the game WoT when you get a larger or smaller alpha. Methods .ceil- to a larger and rounds to the nearest ten. And method.floor- rounds the value to the nearest integer.